Refactor parking lot REST API client, convert more REST API tests
This tidies up the previous commit a little by converting more
REST API tests. It also renames the `parking_rest_helper` module to
the more general 'clients' (shorter, and allows for other clients
to be stored in the same module which is more pythonic) lastly
it restricts some of the checking done by the client itself to
the minimum needed.